+++ aspnetcore/fundamentals/logging/index.md
@@ -923,9 +923,9 @@ The `ILoggerProvider` creates one or more `ILogger` instances. The `ILogger` ins
 The sample:
 
 * Is designed to be a very basic sample that sets the color of the log console by event ID and log level. Loggers generally don't change by event ID and are not specific to log level.
-* Creates different colored console entries per log level and event ID using the following configuration type:
+* Creates different color console entries per log level and event ID using the following configuration type:
 
-[!code-csharp[](index/samples/3.x/CustomLogger/ColoredConsoleLogger/ColoredConsoleLoggerConfiguration.cs?name=snippet)]
+[!code-csharp[](index/samples/3.x/CustomLogger/ColorConsoleLogger/ColorConsoleLoggerConfiguration.cs?name=snippet)]
 
 The preceding code sets the default level to `Warning` and the color to `Yellow`. If the `EventId` is set to 0, we will log all events.
 
@@ -933,7 +933,7 @@ The preceding code sets the default level to `Warning` and the color to `Yellow`
 
 The `ILogger` implementation category name is typically the logging source. For example, the type where the logger is created:
 
-[!code-csharp[](index/samples/3.x/CustomLogger/ColoredConsoleLogger/ColoredConsoleLogger.cs?name=snippet)]
+[!code-csharp[](index/samples/3.x/CustomLogger/ColorConsoleLogger/ColorConsoleLogger.cs?name=snippet)]
 
 The preceding code:
 
@@ -951,9 +951,9 @@ public bool IsEnabled(LogLevel logLevel)
 
 The `LoggerProvider` is the class that creates the logger instances. Maybe it is not needed to create a logger instance per category, but this makes sense for some Loggers, like NLog or log4net. Doing this you are also able to choose different logging output targets per category if needed:
 
-[!code-csharp[](index/samples/3.x/CustomLogger/ColoredConsoleLogger/ColoredConsoleLoggerProvider.cs?name=snippet)]
+[!code-csharp[](index/samples/3.x/CustomLogger/ColorConsoleLogger/ColorConsoleLoggerProvider.cs?name=snippet)]
 
-In the preceding code, <xref:Microsoft.Build.Logging.LoggerDescription.CreateLogger*> creates a single instance of the `ColoredConsoleLogger` per category name and stores it in the [`ConcurrentDictionary<TKey,TValue>`](/dotnet/api/system.collections.concurrent.concurrentdictionary-2);
+In the preceding code, <xref:Microsoft.Build.Logging.LoggerDescription.CreateLogger*> creates a single instance of the `ColorConsoleLogger` per category name and stores it in the [`ConcurrentDictionary<TKey,TValue>`](/dotnet/api/system.collections.concurrent.concurrentdictionary-2);
 
 ### Usage and registration of the custom logger
 
@@ -963,7 +963,7 @@ Register the logger in the `Startup.Configure`:
 
 For the preceding code, provide at least one extension method for the `ILoggerFactory`:
 
-[!code-csharp[](index/samples/3.x/CustomLogger/ColoredConsoleLogger/ColoredConsoleLoggerExtensions.cs?name=snippet)]
+[!code-csharp[](index/samples/3.x/CustomLogger/ColorConsoleLogger/ColorConsoleLoggerExtensions.cs?name=snippet)]
 
 ## Additional resources
